Fix drugs type.

+++ app/src/main/java/it/unibo/preh_frontend/dialog/InputDialogFragment.kt
@@ -70,8 +70,9 @@ class InputDialogFragment : DialogFragment() {
             else -> ""
         }
         if (dtDrugName != "") {
-            val drug = Drug(dtDrugName,
-                    inputValueEditText.text.toString().toInt(),
+            val drug = Drug(
+                    dtDrugName,
+                    inputValueEditText.text.toString().toDouble(),
                     unitEditText.text.toString(),
                     time)
             RetrofitClient.postDrugs(drug)
